Castle scene cutscene conversion
Converted to the new system the present cutscene in a new version of the scene ("Fan_Retro_Castle_Night_CONVERTED").
Added a transition from "pee" animation directly to "crouch_idle" animation in Malphayt animator.
Created a short dialogue placeholder to test cutscenes faster.
The cutscene is recreated based on the already present scene, but based on the script it's not finished.